Moved Number functions from from within Moviepage into separate module, 'misc' folder
